% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{b2k}
\alias{b2k}
\alias{b2k_to_BP}
\alias{b2k_to_BP-method}
\alias{b2k_to_CE}
\alias{b2k_to_CE-method}
\alias{is_b2k}
\alias{is_b2k-method}
\alias{is_b2k,CalibratedAges-method}
\alias{b2k_to_BP,numeric-method}
\alias{b2k_to_BP,matrix-method}
\alias{b2k_to_BP,array-method}
\alias{b2k_to_CE,numeric-method}
\alias{b2k_to_CE,matrix-method}
\alias{b2k_to_CE,array-method}
\title{Before 2000}
\usage{
b2k_to_BP(object)

b2k_to_CE(object)

is_b2k(object)

\S4method{is_b2k}{CalibratedAges}(object)

\S4method{b2k_to_BP}{numeric}(object)

\S4method{b2k_to_BP}{matrix}(object)

\S4method{b2k_to_BP}{array}(object)

\S4method{b2k_to_CE}{numeric}(object)

\S4method{b2k_to_CE}{matrix}(object)

\S4method{b2k_to_CE}{array}(object)
}
\arguments{
\item{object}{An object.}
}
\value{
\itemize{
\item \code{b2k_to_CE()} and \code{b2k_to_BP()} return an object of the same sort as
\code{object} with a new time scale.
\item \code{is_b2k()} returns a \code{\link{logical}} scalar.
}
}
\description{
Converts b2k (before 2000) and between BP (Before Present) or CE (Common Era)
time scales.
}
\note{
There is no year \eqn{0} in BCE/CE scale.
}
\examples{
## BP to CE
BP_to_CE(0)
BP_to_CE(1950)

## BP to b2k
BP_to_b2k(0)
BP_to_b2k(1950)

## CE to BP
CE_to_BP(1950)
CE_to_BP(-1)

## CE to b2k
CE_to_b2k(2000)
CE_to_b2k(-1)

## b2k to BP
b2k_to_BP(0)
b2k_to_BP(50)

## b2k to CE
b2k_to_CE(0)
b2k_to_CE(2000)
}
\seealso{
Other time scales: 
\code{\link{BP}},
\code{\link{CE}}
}
\author{
N. Frerebeau
}
\concept{time scales}
