% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/c14_count.R
\docType{methods}
\name{c14_count}
\alias{c14_count}
\alias{c14_count-method}
\alias{c14_count,CalibratedAges-method}
\title{Radiocarbon Event Count}
\usage{
c14_count(object, ...)

\S4method{c14_count}{CalibratedAges}(
  object,
  from = NULL,
  to = NULL,
  by = 10,
  n = 100,
  calendar = BP(),
  progress = getOption("ananke.progress")
)
}
\arguments{
\item{object}{A \code{\link[=CalibratedAges-class]{CalibratedAges}} object.}

\item{...}{Currently not used.}

\item{from}{length-one \code{\link{numeric}} vector specifying the earliest data to
calibrate for (in cal BP years).}

\item{to}{A length-one \code{\link{numeric}} vector specifying the latest data to
calibrate for (in cal BP years).}

\item{by}{A length-one \code{\link{numeric}} vector specifying the temporal
resolution (in years) of the calibration.}

\item{n}{An \code{\link{integer}} specifying the number of item to choose randomly.}

\item{calendar}{An \code{\link[aion:TimeScale-class]{aion::TimeScale}} object specifying the target
calendar (see \code{\link[aion:calendar]{aion::calendar()}}). Defaults to \code{\link[aion:CE]{aion::CE()}}. If \code{NULL},
\emph{rata die} are returned.}

\item{progress}{A \code{\link{logical}} scalar: should a progress bar be displayed?}
}
\value{
An \code{\link[=RECE-class]{RECE}} object.
}
\description{
Radiocarbon Event Count
}
\note{
This function is currently \emph{experimental}.
}
\examples{
\donttest{
## Radiocarbon data from Bosch et al. 2015
data("ksarakil")

## Calibrate
cal <- c14_calibrate(
  values = ksarakil$date,
  errors = ksarakil$error,
  names = ksarakil$code,
  curves = "marine13",
  reservoir_offsets = 53,
  reservoir_errors = 43,
  from = 50000, to = 0
)

## RECE
tmp <- c14_count(cal, n = 500)
plot(tmp, col = grDevices::hcl.colors(12, "Viridis", rev = TRUE))
}
}
\references{
Carleton, W. C. (2021). Evaluating Bayesian Radiocarbon‐dated Event Count
(REC) Models for the Study of Long‐term Human and Environmental Processes.
\emph{Journal of Quaternary Science}, 36(1): 110‑23. \doi{10.1002/jqs.3256}.
}
\seealso{
Other radiocarbon tools:
\code{\link{F14C}},
\code{\link[=c14_calibrate]{c14_calibrate()}},
\code{\link[=c14_combine]{c14_combine()}},
\code{\link[=c14_curve]{c14_curve()}},
\code{\link{c14_plot}},
\code{\link[=c14_sample]{c14_sample()}},
\code{\link[=c14_spd]{c14_spd()}},
\code{\link[=c14_uncalibrate]{c14_uncalibrate()}},
\code{\link{rec_plot}},
\code{\link[=ridgelines]{ridgelines()}}
}
\author{
N. Frerebeau
}
\concept{radiocarbon tools}
