% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{get_curve}
\alias{get_curve}
\alias{get_curve-method}
\alias{get_curve,character-method}
\title{14C Calibration Curve}
\usage{
get_curve(x, ...)

\S4method{get_curve}{character}(x)
}
\arguments{
\item{x}{A \code{\link{character}} string naming a calibration curve}

\item{...}{Currently not used.}
}
\value{
A three-column \code{\link{data.frame}}: calibrated age BP, uncalibrated age BP and
standard deviation.
}
\description{
14C Calibration Curve
}
\examples{
## Calibrate a single date
calBP <- calibrate(130, 20)
plot(calBP)

## Convert BP scale to CE
calCE <- BP_to_CE(calBP)
plot(calCE)

## Calibrate multiple dates
cal <- calibrate(
  ages = c(5000, 4500),
  errors = c(45, 35),
  names = c("X", "Y")
)
plot(cal, panel.first = graphics::grid())
}
\seealso{
Other radiocarbon tools: 
\code{\link{F14C}},
\code{\link{calibrate}()},
\code{\link{plot_14C}}
}
\author{
N. Frerebeau
}
\concept{radiocarbon tools}
