% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{labels}
\alias{labels}
\alias{labels,CalibratedAges-method}
\title{Find Labels from Object}
\usage{
\S4method{labels}{CalibratedAges}(object, ...)
}
\arguments{
\item{object}{An object from which to find labels.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{character}} vector.
}
\description{
Find a suitable set of labels from an object for use in printing or plotting,
for example.
}
\seealso{
Other mutators:
\code{\link{as.data.frame}},
\code{\link[=as.list]{as.list()}},
\code{\link{mutators}},
\code{\link{subset}}
}
\author{
N. Frerebeau
}
\concept{mutators}
