% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxcal_execute}
\alias{oxcal_execute}
\title{Execute an Oxcal Script}
\usage{
oxcal_execute(script, file = NULL, timeout = 0)
}
\arguments{
\item{script}{A \code{\link{character}} string of instructions for OxCal.}

\item{file}{A \code{\link{character}} string naming a file (without extension) to
write \code{script} to. Output files will be named after \code{file} and written to
the same directory.}

\item{timeout}{An \code{\link{integer}} value specifying the timeout in seconds,
ignored if 0. This is a limit for the elapsed time running OxCal.
Fractions of seconds are ignored (see \code{\link[=system2]{system2()}}).}
}
\value{
Invisibly returns a \code{\link{list}} of paths to the \code{output} files.
}
\description{
Execute an Oxcal Script
}
\note{
Adapted from \code{\link[oxcAAR:executeOxcalScript]{oxcAAR::executeOxcalScript()}}.
}
\examples{
\dontrun{
## Download OxCal
oxcal_setup()

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

plot(res)
plot(res, likelihood = TRUE, posterior = FALSE)
plot(res, likelihood = FALSE, posterior = TRUE)
}
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_analysis_file.html}
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_calibrate}()},
\code{\link{oxcal_download}()},
\code{\link{oxcal_parse}()},
\code{\link{oxcal_setup}()}
}
\author{
M. Hinz, N. Frerebeau
}
\concept{OxCal tools}
