% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/radialplot.R
\docType{methods}
\name{radialplot}
\alias{radialplot}
\alias{radialplot,numeric,numeric-method}
\alias{radialplot,CalibratedAges,missing-method}
\title{Radial Plot}
\usage{
radialplot(values, errors, ...)

\S4method{radialplot}{numeric,numeric}(
  values,
  errors,
  log = TRUE,
  centrality = c("weighted.mean", "mean", "median"),
  bar = FALSE,
  grid = TRUE,
  main = NULL,
  sub = NULL,
  ann = graphics::par("ann"),
  ...
)

\S4method{radialplot}{CalibratedAges,missing}(
  values,
  log = TRUE,
  centrality = c("weighted.mean", "mean", "median"),
  bar = FALSE,
  grid = TRUE,
  main = NULL,
  sub = NULL,
  ann = graphics::par("ann"),
  ...
)
}
\arguments{
\item{values}{A \code{\link{numeric}} vector giving the BP ages or F14C values to be
calibrated (conventional ages) or a \code{\link[=CalibratedAges-class]{CalibratedAges}} object.}

\item{errors}{A \code{\link{numeric}} vector giving the errors associated to the
values to be calibrated.}

\item{...}{Further graphical parameters to be passed to \code{\link[graphics:points]{graphics::points()}}.}

\item{log}{A \code{\link{logical}} scalar: should the z-axis be displayed in
logarithmic scale?}

\item{centrality}{A \code{\link{character}} string specifying a measure of centrality,
used for centering the plot. It must be one of "\code{weighted.mean}" (the
default), "\code{mean}" or "\code{median}". Any unambiguous substring can be given.}

\item{bar}{A \code{\link{logical}} scalar: should a bar showing the 2-sigma range
around the central value be displayed?}

\item{grid}{A \code{\link{logical}} scalar: should a grid (originating at \eqn{[0,0]}
and stretching to the z-scale) be displayed?}

\item{main}{A \code{\link{character}} string giving a main title for the plot.}

\item{sub}{A \code{\link{character}} string giving a subtitle for the plot.}

\item{ann}{A \code{\link{logical}} scalar: should the default annotation (title and x
and y labels) appear on the plot?}
}
\value{
\code{radialplot()} is called it for its side-effects: it results in a graphic
being displayed.
}
\description{
Radial Plot
}
\examples{
## Data from Bosch et al. 2015
data("ksarakil")

## Radialplot
radialplot(ksarakil$date, ksarakil$error, pch = 16)

## Linear z-scale
radialplot(ksarakil$date, ksarakil$error, log = FALSE, pch = 16)
}
\references{
Galbraith, Rex F. (1988). Graphical Display of Estimates Having Differing
Standard Errors. \emph{Technometrics}, 30(3): 271-281.
\doi{10.1080/00401706.1988.10488400}.

Galbraith, Rex F. (1990). The Radial Plot: Graphical Assessment of Spread
in Ages. \emph{International Journal of Radiation Applications and
Instrumentation. Part D. Nuclear Tracks and Radiation Measurements},
17(3): 207-214. \doi{10.1016/1359-0189(90)90036-W}.

Galbraith, Rex F. (1994). Some Applications of Radial Plots. \emph{Journal of
the American Statistical Association}, 89(428): 1232-1242.
\doi{10.1080/01621459.1994.10476864}.
}
\seealso{
Other statistics:
\code{\link{interval_credible}},
\code{\link{interval_hdr}},
\code{\link[=mean]{mean()}},
\code{\link[=median]{median()}},
\code{\link[=quantile]{quantile()}}
}
\author{
N. Frerebeau
}
\concept{statistics}
