% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\name{rec_image}
\alias{rec_image}
\alias{image,RECE-method}
\alias{image,RECE,missing-method}
\title{Plot a Radiocarbon Event Count Ensemble}
\usage{
\S4method{image}{RECE}(x, calendar = getOption("ananke.calendar"), ...)
}
\arguments{
\item{x}{An \code{\linkS4class{RECE}} object.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the target calendar
(see \code{\link[=calendar]{calendar()}}).}

\item{...}{Further parameters to be passed to \code{\link[graphics:image]{graphics::image()}}.}
}
\value{
\code{image()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
\description{
Plot a Radiocarbon Event Count Ensemble
}
\references{
Carleton, W. C. (2021). Evaluating Bayesian Radiocarbon‐dated Event Count
(REC) Models for the Study of Long‐term Human and Environmental Processes.
\emph{Journal of Quaternary Science}, 36(1): 110‑23. \doi{10.1002/jqs.3256}.
}
\author{
N. Frerebeau
}
\concept{modeling tools}
