% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RECE-class}
\alias{RECE-class}
\alias{.RECE}
\title{Radiocarbon Event Count Ensemble}
\description{
An S4 class to represent a radiocarbon event count ensemble.
}
\note{
This class inherits from \code{\link[aion:TimeSeries-class]{aion::TimeSeries}}.
}
\seealso{
Other classes:
\code{\link{CalibratedAges-class}},
\code{\link{CalibratedIntervals-class}},
\code{\link{CalibratedSPD-class}},
\code{\link{ProxyRecord-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
