% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{era}
\alias{era_label}
\alias{era_name}
\alias{era_epoch}
\alias{era_scale}
\alias{era_direction}
\alias{era_unit}
\alias{era_days}
\alias{start}
\alias{time}
\alias{project}
\alias{hpdi}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{chronos}{\code{\link[chronos]{era}}, \code{\link[chronos:era_get]{era_days}}, \code{\link[chronos:era_get]{era_direction}}, \code{\link[chronos:era_get]{era_epoch}}, \code{\link[chronos:era_get]{era_label}}, \code{\link[chronos:era_get]{era_name}}, \code{\link[chronos:era_get]{era_scale}}, \code{\link[chronos:era_get]{era_unit}}, \code{\link[chronos]{hpdi}}, \code{\link[chronos]{project}}, \code{\link[chronos]{start}}, \code{\link[chronos]{time}}}
}}

