% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{ridgelines}
\alias{ridgelines}
\alias{ridgelines-method}
\alias{ridgelines,CalibratedAges-method}
\title{Ridgeline Plot of Calibrated Radiocarbon Ages}
\usage{
ridgelines(x, ...)

\S4method{ridgelines}{CalibratedAges}(
  x,
  calendar = get_calendar(),
  interval = c("hdr", "credible"),
  level = 0.954,
  fixed = TRUE,
  decreasing = FALSE,
  col = "grey",
  ...
)
}
\arguments{
\item{x}{A \code{\link[=CalibratedAges-class]{CalibratedAges}} or \code{\link[=CalibratedSPD-class]{CalibratedSPD}} object.}

\item{...}{Further parameters to be passed to \code{\link[aion:plot]{aion::plot()}}.}

\item{calendar}{An \code{\link[aion:TimeScale-class]{aion::TimeScale}} object specifying the target
calendar (see \code{\link[aion:calendar]{aion::calendar()}}). If \code{NULL}, \emph{rata die} are returned.}

\item{interval}{A \code{\link{character}} string specifying the intervals to be drawn.
It must be one of "\code{hrd}" (the default) or "\code{credible}".
Any unambiguous substring can be given.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.
Only used if \code{interval} is \code{TRUE}.}

\item{fixed}{A \code{\link{logical}} scalar: should a fixed y scale be used?
If \code{TRUE} (the default), ages are equally spaced along the y axis and
arranged in chronological order. If \code{FALSE}, age \code{positions} are used
(see \code{\link[=c14_calibrate]{c14_calibrate()}}).}

\item{decreasing}{A \code{\link{logical}} scalar: should the sort order be decreasing?}

\item{col}{A specification for the plotting colors.}
}
\value{
\code{ridgelines()} is called it for its side-effects: it results in a graphic
being displayed (invisibly returns \code{x}).
}
\description{
Ridgeline Plot of Calibrated Radiocarbon Ages
}
\examples{
## Calibrate multiple dates
cal <- c14_calibrate(
  values = c(5000, 4500),
  errors = c(45, 35),
  names = c("X", "Y")
)

## Ridgelines plot
ridgelines(cal, panel.first = graphics::grid())

## Change colors
ridgelines(cal, col = c("red", "blue"))
}
\seealso{
Other radiocarbon tools:
\code{\link{F14C}},
\code{\link[=c14_calibrate]{c14_calibrate()}},
\code{\link[=c14_combine]{c14_combine()}},
\code{\link[=c14_count]{c14_count()}},
\code{\link[=c14_curve]{c14_curve()}},
\code{\link{c14_plot}},
\code{\link[=c14_sample]{c14_sample()}},
\code{\link[=c14_spd]{c14_spd()}},
\code{\link[=c14_uncalibrate]{c14_uncalibrate()}},
\code{\link{rec_plot}}
}
\author{
N. Frerebeau
}
\concept{radiocarbon tools}
