% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,CalibratedAges-method}
\title{Object Summaries}
\usage{
\S4method{summary}{CalibratedAges}(object, ..., digits = NULL, calendar = get_calendar())
}
\arguments{
\item{object}{A \code{\link[=CalibratedAges-class]{CalibratedAges}} object.}

\item{...}{Currently not used.}

\item{digits}{An \code{\link{integer}} (or \code{NULL}) used for number formatting (see
\code{\link[base:format]{base::format()}}).}

\item{calendar}{An \code{\link[aion:TimeScale-class]{aion::TimeScale}} object specifying the target
calendar (see \code{\link[aion:calendar]{aion::calendar()}}). If \code{NULL}, \emph{rata die} are returned.}
}
\value{
An object of class \code{\link[base:table]{table}}.
}
\description{
Object Summaries
}
\examples{
## Calibrate multiple dates
cal <- c14_calibrate(
  values = c(5000, 4500),
  errors = c(45, 35),
  names = c("X", "Y")
)

## Summary
summary(cal, calendar = CE())
}
\seealso{
Other summary:
\code{\link{describe}}
}
\author{
N. Frerebeau
}
\concept{summary}
