% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ananke-package.R
\docType{package}
\name{ananke-package}
\alias{ananke-package}
\alias{_PACKAGE}
\alias{ananke}
\title{ananke: Analysis of Time Intervals and Temporal Reasoning}
\description{
A toolkit for temporal reasoning. This package allows to describe and analyze finite time intervals. It includes tools for radiocarbon calibration and Bayesian chronological modeling. It also provides classes and methods for the representation of archaeological sequence diagrams in the framework of Dye and Buck (2015) \doi{10.1016/j.jas.2015.08.008}.
}
\details{
\tabular{ll}{
\strong{Package:} \tab ananke \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\section{Package options}{

\code{ananke} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{ananke.progress}: a \code{\link{logical}} scalar. Should progress bars be
displayed?
\item \code{ananke.oxcal}: a \code{\link{character}} string specifying the path to the OxCal
executable.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/ananke/}
  \item \url{https://github.com/tesselle/ananke}
  \item Report bugs at \url{https://github.com/tesselle/ananke/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
33607 Pessac cedex\cr
France
}
\keyword{internal}
