% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{c14_plot}
\alias{c14_plot}
\alias{plot,CalibratedAges,missing-method}
\title{Plot Calibrated Radiocarbon Ages}
\usage{
\S4method{plot}{CalibratedAges,missing}(
  x,
  interval = TRUE,
  level = 0.95,
  decreasing = TRUE,
  main = NULL,
  sub = NULL,
  ann = graphics::par("ann"),
  axes = TRUE,
  frame.plot = FALSE,
  panel.first = NULL,
  panel.last = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{\linkS4class{CalibratedAges}} object.}

\item{interval}{A \code{\link{logical}} scalar: should highest posterior density
interval be drawn?}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{decreasing}{A \code{\link{logical}} scalar: should the sort order be decreasing?}

\item{main}{A \code{\link{character}} string giving a main title for the plot.}

\item{sub}{A \code{\link{character}} string giving a subtitle for the plot.}

\item{ann}{A \code{\link{logical}} scalar: should the default annotation (title and x,
y and z axis labels) appear on the plot?}

\item{axes}{A \code{\link{logical}} scalar: should axes be drawn on the plot?}

\item{frame.plot}{A \code{\link{logical}} scalar: should a box be drawn around the
plot?}

\item{panel.first}{An an \code{expression} to be evaluated after the plot axes are
set up but before any plotting takes place. This can be useful for drawing
background grids.}

\item{panel.last}{An \code{expression} to be evaluated after plotting has taken
place but before the axes, title and box are added.}

\item{...}{Other \link[graphics:par]{graphical parameters} may also be passed as
arguments to this function.}
}
\value{
\code{plot()} is called it for its side-effects: it results in a graphic
being displayed. Invisibly returns \code{x}.
}
\description{
Plot Calibrated Radiocarbon Ages
}
\examples{
## Calibrate a single date
cal <- c14_calibrate(130, 20)
plot(cal)

## Calibrate multiple dates
BP <- c14_calibrate(
  ages = c(5000, 4500),
  errors = c(45, 35),
  names = c("X", "Y")
)
plot(BP, panel.first = graphics::grid())

## Convert BP scale to CE
CE <- BP_to_CE(BP)
plot(CE)

## HPDI
hpdi(BP, level = 0.954)
hpdi(CE, level = 0.954)
}
\seealso{
Other radiocarbon tools: 
\code{\link{F14C}},
\code{\link{c14_calibrate}()},
\code{\link{c14_combine}()},
\code{\link{c14_curve}()}
}
\author{
N. Frerebeau
}
\concept{radiocarbon tools}
