% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/c14_validate.R
\docType{methods}
\name{c14_validate}
\alias{c14_validate}
\alias{c14_validate-method}
\alias{c14_validate,CalibratedAges-method}
\title{Check Calibrated Radiocarbon Dates}
\usage{
c14_validate(object, ...)

\S4method{c14_validate}{CalibratedAges}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{CalibratedAges}} object.}

\item{...}{Currently not used.}
}
\value{
\code{c14_validate()} is called it for its side-effects: it prints
\link[=warnings]{warning messages} if calibrated agea are (partially) out of
calibration range. Invisibly returns \code{x}.
}
\description{
Check Calibrated Radiocarbon Dates
}
\examples{
## Calibrate a single date
cal <- c14_calibrate(300, 20)
plot(cal, panel.first = graphics::grid())

## Calibrate multiple dates
cal <- c14_calibrate(
  values = c(5000, 4500),
  errors = c(45, 35),
  names = c("X", "Y")
)
plot(cal, calendar = BP(), panel.first = graphics::grid())
plot(cal, interval = FALSE)
plot(cal[, 1, ], col.interval = "red")

plot(cal, density = FALSE, level = 0.68, lwd = 5)
plot(cal, density = FALSE, level = 0.95, lwd = 5)

\donttest{
## Out of 14C range?
out <- c14_calibrate(130, 20)
plot(out)
}
}
\seealso{
Other radiocarbon tools: 
\code{\link{F14C}},
\code{\link{c14_calibrate}()},
\code{\link{c14_combine}()},
\code{\link{c14_curve}()},
\code{\link{c14_ensemble}()},
\code{\link{c14_plot}},
\code{\link{c14_spd}()},
\code{\link{c14_uncalibrate}()},
\code{\link{rec_plot}}
}
\author{
N. Frerebeau
}
\concept{radiocarbon tools}
