% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CalibratedAges-class}
\alias{CalibratedAges-class}
\alias{.CalibratedAges}
\title{Calibrated Radiocarbon Ages}
\arguments{
\item{labels}{A \code{\link{character}} vector specifying the names of the ages (e.g.
laboratory codes).}

\item{ages}{A \code{\link{numeric}} vector giving the BP ages to be calibrated.}

\item{errors}{A \code{\link{numeric}} vector giving the standard deviation of the ages
to be calibrated.}

\item{curves}{A \code{\link{character}} vector specifying the calibration curves
used.}
}
\description{
An S4 class to represent calibrated radiocarbon ages.
}
\section{Slots}{

\describe{
\item{\code{start}}{A length-one \code{\link{numeric}} vector specifying the beginning of the
calibrated time range.}

\item{\code{resolution}}{A length-one \code{\link{numeric}} vector specifying the resolution of
the calibrated time range.}

\item{\code{F14C}}{A \code{\link{logical}} scalar:}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{CE}", "\code{BP}" or "\code{b2k}").}
}}

\note{
This class inherits from \code{\link{matrix}}.
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
