% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CalibratedIntervals-class}
\alias{CalibratedIntervals-class}
\alias{.CalibratedIntervals}
\title{Calibrated Intervals}
\description{
An S4 class to represent calibrated intervals of radiocarbon ages.
}
\section{Slots}{

\describe{
\item{\code{p}}{A \code{\link{numeric}} vector giving the probabilities.}
}}

\note{
This class inherits from \code{\link[aion:TimeIntervals-class]{aion::TimeIntervals}}.
}
\seealso{
Other classes: 
\code{\link{CalibratedAges-class}},
\code{\link{CalibratedSPD-class}},
\code{\link{ProxyRecord-class}},
\code{\link{RECE-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
