% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CalibratedSPD-class}
\alias{CalibratedSPD-class}
\alias{.CalibratedSPD}
\title{Calibrated SPD}
\description{
An S4 class to represent summed probability distributions (SPD) of
calibrated radiocarbon ages.
}
\note{
This class inherits from \code{\link[aion:TimeSeries-class]{aion::TimeSeries}}.
}
\seealso{
Other classes: 
\code{\link{CalibratedAges-class}},
\code{\link{CalibratedIntervals-class}},
\code{\link{ProxyRecord-class}},
\code{\link{RECE-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
