% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{get_curve}
\alias{get_curve}
\alias{get_curve-method}
\alias{get_curve,character-method}
\title{14C Calibration Curve}
\usage{
get_curve(x, ...)

\S4method{get_curve}{character}(x)
}
\arguments{
\item{x}{A \code{\link{character}} string naming a calibration curve}

\item{...}{Currently not used.}
}
\value{
A three-column \code{\link{data.frame}}: calibrated age BP, uncalibrated age BP and
standard deviation.
}
\description{
14C Calibration Curve
}
\examples{

}
\seealso{
Other radiocarbon tools: 
\code{\link{F14C}},
\code{\link{calibrate}()}
}
\author{
N. Frerebeau
}
\concept{radiocarbon tools}
