% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{mean}
\alias{mean}
\alias{mean,CalibratedAges-method}
\alias{mean,CalibratedAges,missing-method}
\title{Mean}
\usage{
\S4method{mean}{CalibratedAges}(x, na.rm = FALSE, ..., calendar = getOption("ananke.calendar"))
}
\arguments{
\item{x}{A \code{\linkS4class{CalibratedAges}} object.}

\item{na.rm}{A \code{\link{logical}} scalar: should \code{NA} values be stripped before the
computation proceeds?}

\item{...}{Currently not used.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the target calendar
(see \code{\link[=calendar]{calendar()}}).}
}
\description{
Mean
}
\examples{
## Calibrate a single date
cal <- c14_calibrate(300, 20)

## Statistics
median(cal)
mean(cal)

## Plot
plot(cal, calendar = CE(), panel.first = graphics::grid())

## Need to set 'calendar'
abline(v = median(cal, calendar = CE()), lty = 2, col = "blue")
abline(v = mean(cal, calendar = CE()), lty = 2, col = "red")
}
\seealso{
Other statistics: 
\code{\link{interval_hdr}()},
\code{\link{median}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
