% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{median}
\alias{median}
\alias{median,CalibratedAges-method}
\alias{median,CalibratedAges,missing-method}
\title{Median}
\usage{
\S4method{median}{CalibratedAges}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{CalibratedAges}} object.}

\item{na.rm}{A \code{\link{logical}} scalar: should NA values be stripped before the
computation proceeds?}

\item{...}{Currently not used.}
}
\description{
Median
}
\examples{
## Calibrate a single date
cal <- c14_calibrate(130, 20)
plot(cal)

## Calibrate multiple dates
BP <- c14_calibrate(
  ages = c(5000, 4500),
  errors = c(45, 35),
  names = c("X", "Y")
)
plot(BP, panel.first = graphics::grid())

## Convert BP scale to CE
CE <- BP_to_CE(BP)
plot(CE)

## HPDI
hpdi(BP, level = 0.954)
hpdi(CE, level = 0.954)
}
\seealso{
Other statistics: 
\code{\link{hpdi}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
