% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{median}
\alias{median}
\alias{median,CalibratedAges-method}
\alias{median,CalibratedAges,missing-method}
\title{Median}
\usage{
\S4method{median}{CalibratedAges}(x, na.rm = FALSE, ..., calendar = getOption("ananke.calendar"))
}
\arguments{
\item{x}{A \code{\linkS4class{CalibratedAges}} object.}

\item{na.rm}{A \code{\link{logical}} scalar: should \code{NA} values be stripped before the
computation proceeds?}

\item{...}{Currently not used.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the target calendar
(see \code{\link[=calendar]{calendar()}}).}
}
\description{
Median
}
\examples{
## Calibrate a single date
cal <- c14_calibrate(300, 20)

## Statistics
median(cal)
mean(cal)

## Plot
plot(cal, panel.first = graphics::grid())

## Need to set 'calendar' to NULL
abline(v = median(cal, calendar = NULL), lty = 2, col = "blue")
abline(v = mean(cal, calendar = NULL), lty = 2, col = "red")
}
\seealso{
Other statistics: 
\code{\link{interval_hdr}()},
\code{\link{mean}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
