% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{get}
\alias{set}
\alias{get_calendar}
\alias{get_calendar-method}
\alias{names,CalibratedAges-method}
\alias{names<-,CalibratedAges-method}
\alias{get_calendar,CalibratedAges-method}
\title{Get or Set Parts of an Object}
\usage{
get_calendar(x)

\S4method{names}{CalibratedAges}(x)

\S4method{names}{CalibratedAges}(x) <- value

\S4method{get_calendar}{CalibratedAges}(x)
}
\arguments{
\item{x}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{x}.}
}
\value{
An object of the same sort as \code{x} with the new values assigned.
}
\description{
Getters and setters to extract or replace parts of an object.
}
\author{
N. Frerebeau
}
\concept{mutators}
