% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/proxy.R
\docType{methods}
\name{proxy_ensemble}
\alias{proxy_ensemble}
\alias{proxy_ensemble-method}
\alias{proxy_ensemble,numeric-method}
\title{Layer-Counted Proxy Records Uncertainties}
\usage{
proxy_ensemble(depth, ...)

\S4method{proxy_ensemble}{numeric}(
  depth,
  proxy,
  proxy_error,
  time,
  time_error,
  calendar,
  start = NULL,
  end = NULL,
  by = NULL,
  n = 30,
  progress = getOption("ananke.progress"),
  verbose = getOption("ananke.verbose")
)
}
\arguments{
\item{depth}{A \code{\link{numeric}} vector giving the depth at which proxy values and
calendar ages were measured (in decreasing order).}

\item{...}{Currently not used.}

\item{proxy}{A \code{\link{numeric}} vector giving the proxy values.}

\item{proxy_error}{A \code{\link{numeric}} vector giving the proxy uncertainties.}

\item{time}{A \code{\link{numeric}} vector giving the calendar ages (in years).}

\item{time_error}{A \code{\link{numeric}} vector giving the calendar age uncertainties
(in years).}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the calendar of
\code{time} (see \code{\link[=calendar]{calendar()}}).}

\item{start}{A length-one \code{\link{numeric}} vector specifying the starting value of
the temporal sequence at which densities are to be estimated (in years).}

\item{end}{A length-one \code{\link{numeric}} vector specifying the end value of the
temporal sequence at which densities are to be estimated (in cal BP years).}

\item{by}{A length-one \code{\link{numeric}} vector specifying the increment of
the temporal sequence at which densities are to be estimated (in years).}

\item{n}{An \code{\link{integer}} specifying the number of item to choose.}

\item{progress}{A \code{\link{logical}} scalar: should a progress bar be displayed?}

\item{verbose}{A \code{\link{logical}} scalar: should extra information be reported?}
}
\value{
A \code{\linkS4class{ProxyRecord}} object.
}
\description{
Represents layer-counted proxy records as sequences of probability
distributions on absolute, error-free time axes.
}
\examples{
\dontrun{
## Get NGRIP records
data("ngrip2010", package = "folio")
ngrip2010 <- subset(ngrip2010, !is.na(MCE))
ngrip2010 <- ngrip2010[nrow(ngrip2010):1, ] # Sort in chronological order

## Replicate fig. 3d from Boers et al. (2017)
ngrip_record <- proxy_ensemble(
  depth = ngrip2010$depth,
  proxy = ngrip2010$delta,
  proxy_error = 0.01,
  time = ngrip2010$age,
  time_error = ngrip2010$MCE,
  calendar = b2k(), # /!\
  by = 20,
  n = 30
)

plot(ngrip_record)
}
}
\references{
Boers, N., Goswami, B. & Ghil, M. (2017). A Complete Representation of
Uncertainties in Layer-Counted Paleoclimatic Archives. \emph{Climate of the
Past}, 13(9): 1169-1180. \doi{10.5194/cp-13-1169-2017}.
}
\seealso{
Other modeling tools: 
\code{\link{proxy_plot}}
}
\author{
N. Frerebeau
}
\concept{modeling tools}
