% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/proxy.R
\docType{methods}
\name{proxy_plot}
\alias{proxy_plot}
\alias{plot,ProxyRecord,missing-method}
\title{Plot Layer-Counted Proxy Records Uncertainties}
\usage{
\S4method{plot}{ProxyRecord,missing}(x, calendar = getOption("ananke.calendar"), ...)
}
\arguments{
\item{x}{A \code{\linkS4class{ProxyRecord}} object.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the target calendar
(see \code{\link[=calendar]{calendar()}}).}

\item{...}{Further parameters to be passed to \code{\link[graphics:image]{graphics::image()}}.}
}
\value{
\code{plot()} is called it for its side-effects: it results in a graphic
being displayed. Invisibly returns \code{x}.
}
\description{
Plot Layer-Counted Proxy Records Uncertainties
}
\examples{
\dontrun{
## Get NGRIP records
data("ngrip2010", package = "folio")
ngrip2010 <- subset(ngrip2010, !is.na(MCE))
ngrip2010 <- ngrip2010[nrow(ngrip2010):1, ] # Sort in chronological order

## Replicate fig. 3d from Boers et al. (2017)
ngrip_record <- proxy_ensemble(
  depth = ngrip2010$depth,
  proxy = ngrip2010$delta,
  proxy_error = 0.01,
  time = ngrip2010$age,
  time_error = ngrip2010$MCE,
  calendar = b2k(), # /!\
  by = 20,
  n = 30
)

plot(ngrip_record)
}
}
\seealso{
Other proxy tools: 
\code{\link{proxy_ensemble}()}
}
\author{
N. Frerebeau
}
\concept{proxy tools}
