% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ProxyRecord-class}
\alias{ProxyRecord-class}
\alias{.ProxyRecord}
\title{Proxy Record}
\description{
An S4 class to store proxy records.
}
\section{Slots}{

\describe{
\item{\code{year}}{A \code{\link{numeric}} vector giving the time points at which the
distribution is estimated.}
}}

\note{
This class inherits from \code{\linkS4class{TimeSeries}}.
}
\seealso{
Other classes: 
\code{\link{CalibratedAges-class}},
\code{\link{CalibratedSPD-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
