% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ProxyRecord-class}
\alias{ProxyRecord-class}
\alias{.ProxyRecord}
\title{Proxy Record}
\description{
An S4 class to store proxy records.
}
\note{
This class inherits from \code{\link[aion:TimeSeries-class]{aion::TimeSeries}}.
}
\seealso{
Other classes: 
\code{\link{CalibratedAges-class}},
\code{\link{CalibratedIntervals-class}},
\code{\link{CalibratedSPD-class}},
\code{\link{RECE-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
