% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{quantile}
\alias{quantile}
\alias{quantile,CalibratedAges-method}
\alias{quantile,CalibratedAges,missing-method}
\alias{quantile,ProxyRecord-method}
\alias{quantile,ProxyRecord,missing-method}
\title{Quantiles of a Density Estimate}
\usage{
\S4method{quantile}{CalibratedAges}(
  x,
  probs = seq(0, 1, 0.25),
  na.rm = FALSE,
  ...,
  calendar = get_calendar()
)

\S4method{quantile}{ProxyRecord}(x, probs = seq(0, 1, 0.25), na.rm = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[=CalibratedAges-class]{CalibratedAges}} object.}

\item{probs}{A \code{\link{numeric}} vector of probabilities with values in
\eqn{[0,1]}.}

\item{na.rm}{A \code{\link{logical}} scalar: should \code{NA} values be stripped before the
computation proceeds?}

\item{...}{Currently not used.}

\item{calendar}{An \code{\link[aion:TimeScale-class]{aion::TimeScale}} object specifying the target
calendar (see \code{\link[aion:calendar]{aion::calendar()}}). If \code{NULL}, \emph{rata die} are returned.}
}
\value{
A \code{numeric} \code{\link{matrix}} containing the quantiles.
}
\description{
Quantiles of a Density Estimate
}
\examples{
## Calibrate multiple dates
cal <- c14_calibrate(
  values = c(5000, 4500),
  errors = c(45, 35),
  names = c("X", "Y")
)

## Statistics
quantile(cal)
median(cal)
mean(cal)

## Plot
plot(cal, calendar = CE())

## Need to set 'calendar'
abline(v = median(cal, calendar = CE()), lty = 2, col = "blue")
abline(v = mean(cal, calendar = CE()), lty = 2, col = "red")
}
\seealso{
Other statistics:
\code{\link{interval_credible}},
\code{\link{interval_hdr}},
\code{\link[=mean]{mean()}},
\code{\link[=median]{median()}},
\code{\link[=radialplot]{radialplot()}}
}
\author{
N. Frerebeau
}
\concept{statistics}
